+++ workers/unity/Assets/Fps/Scripts/SetupLogic/WorkerConnectorBase.cs
@@ -1,3 +1,5 @@
+using System;
+using System.Collections;
 using System.Threading.Tasks;
 using Improbable.Gdk.Core;
 using Improbable.Worker.CInterop;
@@ -11,7 +13,7 @@ namespace Fps
 
         [SerializeField] protected MapBuilderSettings MapBuilderSettings;
 
-        protected GameObject levelInstance;
+        [NonSerialized] internal GameObject LevelInstance;
 
         protected abstract string GetWorkerType();
 
@@ -34,29 +36,30 @@ namespace Fps
 
         protected override void HandleWorkerConnectionEstablished()
         {
-            LoadWorld();
+            StartCoroutine(LoadWorld());
         }
 
         public override void Dispose()
         {
-            if (levelInstance != null)
+            if (LevelInstance != null)
             {
-                Destroy(levelInstance);
-                levelInstance = null;
+                Destroy(LevelInstance);
+                LevelInstance = null;
             }
 
             base.Dispose();
         }
 
         // Get the world size from the config, and use it to generate the correct-sized level
-        protected virtual void LoadWorld()
+        protected virtual IEnumerator LoadWorld()
         {
-            levelInstance = MapBuilder.GenerateMap(
+            yield return MapBuilder.GenerateMap(
                 MapBuilderSettings,
                 transform,
                 Worker.Connection,
                 Worker.WorkerType,
-                Worker.LogDispatcher);
+                Worker.LogDispatcher,
+                this);
         }
     }
 }